* EnglishPatient 플러그인이 글도 필터링 * Mass deleting 실행 전에 확인 창 추가 * SimpleSpam 플러그인 우선순위 조정

+++ actions/post/edit.php
@@ -12,7 +12,7 @@ if (is_post()) {
 
 	unset($_POST['post']['password']);
 	$post->import($_POST['post']);
-	if ($_POST['delete']) {
+	if ($board->use_attachment && isset($_POST['delete'])) {
 		foreach ($_POST['delete'] as $id) {
 			$attachment = Attachment::find($id);
 			$attachment->delete();
